refactor: create a component to display rarity value as stars
The style is now unified in all the places where rarity was being displayed, although the spec was undefined for some of those contexts.

N.B. using unary + to turn a string into the apparent number is actually documented and allegedly expected behavior. Should be stable as a temp fix, tagged it with the associated issue so we can remove it once the server is up to date.