refactor: create a component to display rarity value as stars (#7)
* refactor: create a component to display rarity value as stars

The style is now unified in all the places where rarity was being displayed, although the spec was undefined for some of those contexts.

N.B. using unary + to turn a string into the apparent number is actually documented and allegedly expected behavior. Should be stable as a temp fix, tagged it with the associated issue so we can remove it once the server is up to date.

* style: fileoverview had ambiguous wording for multiplier pipe

* use math.max for the multiplier pipe

This also negates the need for the unary as the rarity can be interpreted as a number at this stage ... for some reason (i.e. Math.max("5", 1) = 5). :|