refactor: weapon list page ft. css grid (#16)
* Add constants for gradient bgs + level 1 module color

* Create globalized styles for list pages

In the very least they'll share host, header, and content stylings. Grid may depend on the content though, so we'll hold off on that until we've got more finalized stylings for weapon and artifact cards.

* First pass on weapons page refactor

Simplify DOM, apply grid, and componentize the weapons cards. From here, I'd like to experiment with some different formats for the weapon card displays so that 1) we don't mistakenly think they're generalizable with the character cards and 2) it'll easily transition into the filter / sort feature.

* Generalize cover element style

Will be a common approach for applying links over cards

* Give weapon names full bleed on weapon cards

Some of them are ridiculously long so let's give them extra room to grow.

* Add substat to weapon cards

Looks like some of the underlying data is a bit noisy but let's include it anyway so it's easier to see what changes are required.

* Add tests to weapon card

Documenting the current expected behavior

* Use rem in character card stylings

Just to keep it more consistent, since em may have unexpected effects if the contextual font size changes.