Adjust fun constants on the character page
When the region banner is right under the character tabs, the contrast may be too low to see which tab is currently selected. So let's have a slightly different approach where there's a bit more negative space around the banner. Also make sure the image height doesn't force an overflow under the footer.