Impact Angular 2.0 (#172)
* Flatten z-indices on cards

Since these hover links should only appear above the regular content flow, let's reduce this number to 1 instead.

* Simplify DOM on the stats page

* Simplify footer styling

Namely needed to remove the unnecessary z-index, but noticed that the font / link sizing could be more accessible. Updated to more common standard of 12px / 0.75rem + simplified the flex rules so the text wouldn't trigger an overflow on the footer container.

* Similar style treatments to login and signup pages

* Overhaul of page body + nav to simplify sticky header

Made the DOM flatter, header actually sticky now (weird bug where position absolute wasn't working in prod). Simplified some of the code backing the nav and made some of the text bigger. Still need to address some unexpected behaviors on individualized pages.

* Weapons page rework

Part 1: Streamlining the content for mobile. Constrained width for desktop but will probably require a separate treatment; just wanted to get the DOM simplification sorted first.

* Least terrible mobile / desktop reflow I can think of

It's not much, but it's honest work

* Add word break to weapon ability description

Wtf Mistsplitter?

* Remove ids

From a failed mobile anchor scroll experiment. Might revisit later.

* Throw that footer on the GROUND

The bottom of the footer will sit on the bottom of the page when there's not enough content in the body to fill the page.

* Artifacts Rework

Following the same approach as the weapons page, minimizing the display for a better mobile experience and reflowing for desktop.

* Fixed width for weapon desktop side display

Checking on edge cases, should just be a set width in the case that there is no ability or description text to pad out the element.

* Simplify artifacts routing

Just cuz additional tabs aren't implemented yet and there's probably a better API structure to back what it was intended for as well.

* Use ng-templates for loading

Just cuz the additional DOM hierarchy isn't required; we only care about gating the UI on whether the entity was actually loaded.

* Remove stub section for materials on weapons page

* First pass on character page

It's a big 'un so just tackling the header on the initial pass.

* Smaller display for weapon, artifact headers

* Some nits to the HTML stylings

Checking against some of the available data in build notes, just trying to make it look decent lol ...

* Centralize weapon summary viewmodel

Since we'll need to build a minified view off the same smaller set of weapon data, let's move the model into a "viewmodel" folder.

* Create component for minified weapon display esp. for character builds

* Centralize item header for weapons and artifacts

I think it looks right for these pages, so let's make one component that can centralize some of the different CSS decisions etc.

* I tried so hard and got so far

But I might take an ultra simplified approach on this, I just want this set of changes in the history just in case I want to reference it / roll back

* Significantly less fun character page

Stripped down to the bare minimum based on my understanding of the key product goals.

* Highlight selected tab in Angular-esque way

I don't make the rules

* Basic display for talents

* Minor UI adjustment to talent description

* Basic display for constellations

* Reflow into three columns for really big displays

Seems to work a little better like this...

* Margin to talent / constellation img in case of text overflow

* Attempt to incorporate some amount of "fun" colors

I have no soul

* Reflow option a: weapons on left, artifacts + stats on right

* Collapsible build sections

With material icons

* Update toggle UX

* Simplify character data requests

Should probably structure the API a bit differently but this should reduce the number of requests, since they were being refreshed for each tab change rather than cached locally.

* Handle case with no data

Since Aloy has no constellations, let's make sure the display indicates that the request did happen, there was just nothing to display.

* Make sure grid stylings only applied to results

This can be simplified if the API is simplified to a single roundtrip...

* Fun with colors

The accent seems too minimal otherwise ...

* Delegate links to a component

Prefer an abstraction of the data with *ngFor to ensure consistency ...

* Build out preliminary artifact display

Just dumping out the max available info; there are some parts that aren't ready to be hooked up (image, url).

* Adjust CSS in weapon chips

Just giving the title a little bit of padding to differentiate from the attributes section

* Trim down artifact chip based on available data

Since there's no image yet, we can spend more pixels on the descriptions.

* Add separator between artifact set recommendations

* Transform artifact data

It's a bit messy but this should be handled by the server side to begin with ...

* Condensed weapon stats display on chip

Just cuz it's weird to see it scooted over towards the image on larger displays

* Fix character list page

Kind of wonked out in the reworks

* Add isRecommended back in to character builds

Probably another good attribute to compute server side on the build objects directly ...

* Slightly more fun character page

I like the region banner and character art, so let's try to add it back in for desktop displays ...

* Fix tests

Surprisingly most of them still pass wwwww

* Adjust fun constants on the character page

When the region banner is right under the character tabs, the contrast may be too low to see which tab is currently selected. So let's have a slightly different approach where there's a bit more negative space around the banner. Also make sure the image height doesn't force an overflow under the footer.

* Constrain narrow character art

Since I made the display width-based, some of the more narrow art will be really tall. Let's constrain the height so the displays have a little less variance.

* Account for edge case where builds are missing data

* Realized the sticky header glitched out, so now using different method to handle images

It's only a hack if you add a note: this should never go to production.

* Fix padding in mobile / tablet

I think I'm approaching the limit of what I can do in one set of changes ...

* Collapse character builds by default

Might work better to give an overview of the types of builds first, then you can expand the details? It's trivial to flip anyway.

* Fix home page stylings

They got super fricked and I was too scared to look at it but then it wasn't that bad lol

* Organize imports in app.module.ts

* Clean up artifact chips

ArtifactSetSummary object moved into viewmodels, added tests, organized imports

* Hide missing data on artifact page

If there's no lore or description let's not display the containing divs...

* Organize imports on character.component.ts

* Formalize a few magic numbers

* Cleaning up more constants ft. text colors

* Account for a few edge cases on weapons page.

* Run linty

Bang bangs aren't in vogue anymore :(

* Add test to artifact chip

Reviewing me own code and realized I overlooked this particular case

* noopener the footer